+++ main/src/main.js
@@ -1,23 +1,24 @@
 import { Telegraf, Markup, session } from "telegraf";
 import { message } from "telegraf/filters";
-import { downloadFile, mergeAudioFilesToMp3, createVoice, updateNumbersInJson, banUser, unbanUser, slowDownAudioYa } from "./functions.js";
+import { downloadFile, mergeAudioFilesToMp3, createVoice, updateNumbersInJson, banUser, unbanUser } from "./functions.js";
 import config from "config";
 import fs from "fs";
 import path from "path";
 import ffmpeg from 'fluent-ffmpeg';
 import ffmpegInstaller from '@ffmpeg-installer/ffmpeg';
 
-import { showConsole, hideConsole } from "node-hide-console-window";
-
-
 import { INITIAL_SESSION } from "./variables.js"
 import { setBotCommands, registerBotCommands } from "./botCommands.js";
-import { showMenu, processAudioMessage, is_youtube_url, separateAudioBot, sendMessageToAllUsers, printCurrentTime, processVideo, saveSuggestion, processAiCover, Semaphore, protectBot, showCreateMenu, sendMessageToUser, checkForBan, noteOctaveToFrequency } from "./botFunction.js";
+import { showMenu, processAudioMessage, is_youtube_url, separateAudioBot, printCurrentTime, processVideo, saveSuggestion, processAiCover, Semaphore, protectBot, showCreateMenu,checkForBan, noteOctaveToFrequency, createSessionFolder } from "./botFunction.js";
 import { registerBotActions } from "./botActions.js";
 import { downloadFromYoutube } from "./functions.js";
 
-import { handlePredlog, handlePresetSave, handleYoutubeCover, handleSettings, textHandler } from "./handlers.js";
+import { handlePredlog, handlePresetSave, handleYoutubeCover, handleSettings, textHandler, separateHanlder } from "./handlers.js";
 import { generateSpeechYA } from "./yandexTTS.js";
+import { sendMessageToAllUsers,sendMessageToUser } from "./admin/botFunctions.js";
+import { adminHandler } from "./admin/handler.js";
+import { effectHanlder } from "./effects/handler.js";
+import { createModelHanlder } from "./createModel/handler.js";
 
 // Указываем путь к ffmpeg
 ffmpeg.setFfmpegPath(ffmpegInstaller.path);
@@ -91,246 +92,30 @@ bot.on("video", async (ctx) => {
 
 bot.on(message("text"), async (ctx) => {
   // if (protectBot(ctx)) return
-  if(await checkForBan(ctx)) return
-
+  if (await checkForBan(ctx)) return
   try {
+    // Каждый хендлер должен иметь  Promise.resolve(false) для правильной работы
+    const handlersArray = [
+      adminHandler(ctx),
+      effectHanlder(ctx),
+      createModelHanlder(ctx),
+      separateHanlder(ctx),
+      handleSettings(ctx),
+      ctx.session.waitForPredlog ? handlePredlog(ctx) : Promise.resolve(false),
+      ctx.session.waitForPresetSave ? handlePresetSave(ctx) : Promise.resolve(false),
+      is_youtube_url(ctx.message.text) ? handleYoutubeCover(ctx) : Promise.resolve(false),
+    ];
 
-    if(ctx.session.waitForBan){
-      const uniqueId = ctx.message.text; // получаем уникальный идентификатор пользователя
-      banUser(uniqueId)
-      ctx.reply("Пользователь успешно забанен")
-      ctx.session.waitForBan = false
-      return 
+    const results = await Promise.all(handlersArray);
+
+    if (results.every(result => result === false)) {
+      createSessionFolder(ctx);
+      textHandler(ctx);
     }
 
-    if(ctx.session.waitForUnBan){
-      const uniqueId = ctx.message.text; // получаем уникальный идентификатор пользователя
-      unbanUser(uniqueId)
-      ctx.reply("Пользователь успешно забанен")
-      ctx.session.waitForUnBan = false
-      return 
-    }
-
-
-    if(ctx.session.waitForEchoDelay){
-      const input = ctx.message.text; // входные данные пользователя
-      ctx.session.echoDelay = Math.max(0.05, Math.min(3, parseFloat(input))); // проверяем и присваиваем входные данные к echoDelay
-      ctx.session.waitForEchoDelay = false;
-      ctx.reply(`Значение echoDelay было успешно установлено как ${ctx.session.echoDelay}`, Markup.inlineKeyboard([
-          Markup.button.callback('Назад', 'effects_settings')
-      ]));
-      return;
-  }
-  
-  if(ctx.session.waitForEchoPower){
-      const input = ctx.message.text; // входные данные пользователя
-      ctx.session.echoPower = Math.max(0.05, Math.min(3, parseFloat(input))); // проверяем и присваиваем входные данные к echoPower
-      ctx.session.waitForEchoPower = false;
-      ctx.reply(`Значение echoPower было успешно установлено как ${ctx.session.echoPower}`, Markup.inlineKeyboard([
-          Markup.button.callback('Назад', 'effects_settings')
-      ]));
-      return;
-  }
-  
-  if(ctx.session.waitForReverb){
-      const input = ctx.message.text; // входные данные пользователя
-      ctx.session.reverbPower = Math.max(0.00001, Math.min(1, parseFloat(input))); // проверяем и присваиваем входные данные к reverbPower
-      ctx.session.waitForReverb = false;
-      ctx.reply(`Значение reverbPower было успешно установлено как ${ctx.session.reverbPower}`, Markup.inlineKeyboard([
-          Markup.button.callback('Назад', 'effects_settings')
-      ]));
-      return;
-  }
-  
-  if(ctx.session.waitForAutotuneAttack){
-      const input = ctx.message.text; // входные данные пользователя
-      ctx.session.autotune_attack = Math.max(0.05, Math.min(1, parseFloat(input))); // проверяем и присваиваем входные данные к autotune_attack
-      ctx.session.waitForAutotuneAttack = false;
-      ctx.reply(`Значение autotune_attack было успешно установлено как ${ctx.session.autotune_attack}`, Markup.inlineKeyboard([
-          Markup.button.callback('Назад', 'effects_settings')
-      ]));
-      return;
-  }
-  
-  if(ctx.session.waitForAutotuneStr){
-      const input = ctx.message.text; // входные данные пользователя
-      ctx.session.autotune_strength = Math.max(0.05, Math.min(1, parseFloat(input))); // проверяем и присваиваем входные данные к autotune_strength
-      ctx.session.waitForAutotuneStr = false;
-      ctx.reply(`Значение autotune_strength было успешно установлено как ${ctx.session.autotune_strength}`, Markup.inlineKeyboard([
-          Markup.button.callback('Назад', 'effects_settings')
-      ]));
-      return;
-  }
-
-    if(ctx.session.waitForMinPich){
-      const input = ctx.message.text; // ввод пользователя
-    
-      // Проверяем, ввел ли пользователь частоту в Гц (число от 1 до 16000)
-      const hzValue = parseFloat(input);
-      if (hzValue >= 1 && hzValue <= 16000) {
-        // прямо записываем значение в Гц
-        ctx.session.minPich = hzValue;
-        ctx.reply(`Значение в ${ctx.session.minPich} было сохраннено`)
-      } else {
-        // Проверяем, ввел ли пользователь пару [нота][октава]
-        const noteMatch = input.match(/^([A-G]b?#?)(-?\d+)$/i);
-        if (noteMatch) {
-          const note = noteMatch[1];
-          const octave = parseInt(noteMatch[2], 10);
-          // конвертируем [нота][октава] в Гц и записываем
-          ctx.session.minPich = Math.floor(noteOctaveToFrequency(note, octave));
-          ctx.reply(`Значение в ${ctx.session.minPich} было сохраннено`)
-        } else {
-          ctx.reply("Неверный ввод, вводите на английской раскладке или значения от 1 до 16000")
-          // Неверный ввод, можно отправить сообщение об ошибке
-        }
-      }
-
-      showMenu(ctx)
-      ctx.session.waitForMinPich = false
-      return
-    }
-
-    if(ctx.session.waitForMaxPich){
-      const input = ctx.message.text; // ввод пользователя
-    
-      // Проверяем, ввел ли пользователь частоту в Гц (число от 1 до 16000)
-      const hzValue = parseFloat(input);
-      if (hzValue >= 1 && hzValue <= 16000) {
-        // прямо записываем значение в Гц
-        ctx.session.maxPich = hzValue;
-        ctx.reply(`Значение в ${ctx.session.maxPich} было сохраннено`)
-      } else {
-        // Проверяем, ввел ли пользователь пару [нота][окtава]
-        const noteMatch = input.match(/^([A-G]b?#?)(-?\d+)$/i);
-        if (noteMatch) {
-          const note = noteMatch[1];
-          const octave = parseInt(noteMatch[2], 10);
-          // конвертируем [нота][окtава] в Гц и записываем
-          ctx.session.maxPich = Math.floor(noteOctaveToFrequency(note, octave));
-          ctx.reply(`Значение в ${ctx.session.maxPich} было сохраннено`)
-        } else {
-          ctx.reply("Неверный ввод, вводите на английской раскладке или значения от 1 до 16000")
-          // Неверный ввод, можно отправить сообщение об ошибке
-        }
-      }
-
-      showMenu(ctx)
-      ctx.session.waitForMaxPich = false
-      return
-    }
-
-    if(ctx.session.waitForControlPower){
-      ctx.session.waitForControlPower = false
-      const [transform,silero,separate] = ctx.message.text.split(",")
-
-      updateNumbersInJson(transform,silero,separate)
-      ctx.reply("Данные успешно обновленны")
-      return
-    }
-
-    if(ctx.session.waitForAnonceAll){
-      ctx.session.waitForAnonceAll = false
-      sendMessageToAllUsers(ctx.message.text, bot)
-      return
-    }
-
-    if(ctx.session.waitForAnonceCurrent){
-      ctx.session.waitForAnonceCurrent = false
-
-      const [id,message] = ctx.message.text.split(",")
-      sendMessageToUser(id,message, bot)
-      return
-    }
-
-    
-
-    if(ctx.session.waitForModelName){
-      ctx.session.voiceModelName = ctx.message.text
-      await ctx.reply("Ваша модель была названа: "+ctx.session.voiceModelName)
-      await ctx.reply("Теперь вам нужно ввести краткое описание модели")
-
-      ctx.session.waitForModelName = false
-      ctx.session.waitForModelDesc = true
-      return
-    }
-
-    if(ctx.session.waitForModelDesc){
-      ctx.session.voiceModelDesc = ctx.message.text
-      await ctx.reply("Описание для вашей модели было записанно")
-      await ctx.reply("Введите пол модели, обязательно вводите по шаблону\nmale\nfemale\nchild")
-
-      ctx.session.waitForModelDesc = false
-      ctx.session.waitForModelGender = true
-      return
-    }
-
-    if(ctx.session.waitForModelGender){
-      ctx.session.voiceModelGender = ctx.message.text
-      await ctx.reply("Пол для вашей модели был записан")
-  
-      ctx.session.waitForModelGender = false
-      await ctx.reply("Инициализация модели прошла успешно, вы можете добавлять образцы голоса", Markup.inlineKeyboard([
-        Markup.button.callback('Меню', 'show_create_voice_menu')
-      ]))
-      return
-    }
-
-    if (ctx.session.waitForSeparate) {
-      if (!is_youtube_url(ctx.message.text)) {
-        ctx.reply("Неправильная ссылка на ютуб, введите команду /separate и попробуйте еще раз")
-        ctx.session.waitForSeparate = false
-        return
-      }
-      const uniqueId = ctx.from.id; // получаем уникальный идентификатор пользователя
-      const username = ctx.from.username; // получаем ник пользователя
-      const messageId = ctx.message.message_id; // получаем уникальный идентификатор сообщения
-      const sessionPath = `sessions/${uniqueId}/${messageId}`;
-
-      await downloadFromYoutube(ctx.message.text, sessionPath);
-
-      await separateAudioBot(ctx, sessionPath, true)
-      ctx.session.waitForSeparate = false
-      return
-
-    }
-
-    // Обработка сохранения пресета
-    if (ctx.session.waitForPredlog && handlePredlog(ctx)) {
-      return;
-    }
-
-    // Обработка сохранения пресета
-    if (ctx.session.waitForPresetSave && handlePresetSave(ctx)) {
-      return;
-    }
-
-    // Обработка ютуб ссылки для кавера
-    if (is_youtube_url(ctx.message.text) && handleYoutubeCover(ctx)) {
-      return;
-    }
-
-    // Обработка настроек
-    if (await handleSettings(ctx)) {
-      return;
-    }
-    else {
-      const uniqueId = ctx.from.id; // получаем уникальный идентификатор пользователя
-      const username = ctx.from.username; // получаем ник пользователя
-      const messageId = ctx.message.message_id; // получаем уникальный идентификатор сообщения
-      const sessionPath = `sessions/${uniqueId}/${messageId}`;
-
-      // Создаем папку для пользователя, если она еще не существует
-      if (!fs.existsSync(sessionPath)) {
-        fs.mkdirSync(sessionPath, { recursive: true });
-      }
-
-      // await ctx.reply(ctx.message.text)
-      textHandler(ctx)
-    }
   } catch (err) {
-    console.log(err)
-    ctx.reply("Произошла ошибка при обработке сообщения, повторите снова")
+    console.log(err);
+    ctx.reply("Произошла ошибка при обработке сообщения, повторите снова");
   }
 });
 
@@ -581,7 +366,7 @@ bot.on("audio", async (ctx) => {
 bot.launch();
 
 // Restart msg
-// sendMessageToAllUsers("Бот был перезапущен, все настройки сброшенны\nВведите /start для начала работы", bot)
+sendMessageToAllUsers("Бот был перезапущен, все настройки сброшенны\nВведите /start для начала работы", bot)
 // sendMessageToAllUsers("Бот был обновлен, все подробности в информационном канале https://t.me/mister_parodist_info", bot)
 // sendMessageToAllUsers("Бот временно не работает, тех.работы", bot)
